+++ Assets/ViveGrip/Examples/Scripts/ViveGripExample_Lever.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class ViveGripExample_Lever : MonoBehaviour {
   private ViveGrip_ControllerHandler controller;
   private float oldXRotation;
-  private float VIBRATION_DURATION_IN_MILLISECONDS = 50;
+  private int VIBRATION_DURATION_IN_MILLISECONDS = 50;
   private float MAX_VIBRATION_STRENGTH = 0.7f;
   private float MAX_VIBRATION_ANGLE = 35f;