+++ CHANGELOG.md
@@ -3,6 +3,25 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [2.7.0] - 2017-02-20
+### Changed
+- `ViveGrip_GripPoint.controller.Vibrate(int milliseconds, float strength [0.0-1.0])`
+- Minor code improvements
+### Added
+- `ViveGrip_GripPoint.UpdateRadius(float touch, float hold)`
+
+## [2.6.1] - 2017-01-10
+### Fixed
+- `ViveGrip_ControllerHandler.Device()` will return null when the device isn't ready
+
+## [2.6.0] - 2016-11-11
+### Added
+- `ViveGrip_EventBridge` lets you connect methods to Vive Grip events through the Inspector
+- The Switch example to show off the Event Bridge
+- The Extension Cube example to show off extending grabbing and highlighting in code
+### Fixed
+- Core Vive Grip scripts now obey Unity's `enabled` flags
 ### Changed
 - Moved releases to their own directory
 - Tweaked the documentation
@@ -115,7 +134,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Demo scene
 - Documentation
 
-[Unreleased]: https://github.com/JScott/ViveGrip/compare/v2.5.0...HEAD
+[Unreleased]: https://github.com/JScott/ViveGrip/compare/v2.7.0...HEAD
+[2.7.0]: https://github.com/JScott/ViveGrip/compare/v2.6.1...v2.7.0
+[2.6.1]: https://github.com/JScott/ViveGrip/compare/v2.6.0...v2.6.1
+[2.6.0]: https://github.com/JScott/ViveGrip/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/JScott/ViveGrip/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/JScott/ViveGrip/compare/v2.3.0...v2.4.0
 [2.3.0]: https://github.com/JScott/ViveGrip/compare/v2.2.0...v2.3.0